Fixes some slides pushing the whole deck up...
The version 1.1.0 of bespoke-hash introduced hashing by a slide id attribute, as well as
the data-bespoke-hash attribute. As we were using markdownItAnchor plugin, which automatically
adds slugified ids to h1, h2..., if the data-bespoke-hash value used was the same as slide title,
which happened almost all of the times, the browser understood it had to scroll to put the
heading into the view, which made the whole deck move a bit upwards.

This issue took some good 4-5h, as I thought it had relation to the slide layouts.

The fix was to drop the use of markdownItAnchor - which I don't quite remeber the need for it.